+++ packages/document/template/feedback/message-position.vue
@@ -14,7 +14,7 @@
 </template>
 <script lang="ts" setup>
 const textMap: Record<string, string> = {
-	'top-eft': 'Top Left',
+	'top-left': 'Top Left',
 	top: 'Top',
 	'top-right': 'Top Right',
 	'bottom-left': 'Bottom Left',