+++ packages/document/template/feedback/notification-basic.vue
@@ -2,21 +2,22 @@
 	<px-space>
 		<px-button @click="callByWindow">Window</px-button>
 		<px-button @click="callByCurrentInstance">Current Instance</px-button>
-		<px-button @click="callByOnDemandImport">On-demand Import</px-button>
 	</px-space>
 </template>
 <script lang="ts" setup>
 import { getCurrentInstance } from 'vue'
-import { Message } from '@pixelium/web-vue/es'
 const instance = getCurrentInstance()
 
 const callByWindow = () => {
-	$message('Global Message')
+	$notification.notice({
+		title: 'Via global variable',
+		content: '$notification'
+	})
 }
 const callByCurrentInstance = () => {
-	instance?.appContext.config.globalProperties.PixeliumVue.message('Global Message')
-}
-const callByOnDemandImport = () => {
-	Message('Global Message')
+	instance?.appContext.config.globalProperties.PixeliumVue.notification.info({
+		content: 'instance?.appContext.config.globalProperties.PixeliumVue.notification',
+		title: 'Via Instance'
+	})
 }
 </script>